refactor(layout): 홈페이지 및 루트 레이아웃 구조 개선
 RootLayout을 app router 구조로 이동
 HomePage를 app router의 page.tsx로 이동
 Header/MobileNav를 RootNav로 통합하여 네비게이션 구조 개선
 Button 컴포넌트 import 경로를 새로운 구조에 맞게 수정
 Link 컴포넌트 import 경로를 절대 경로로 변경
 MoveToCommunityButton을 ButtonLink로 변경하여 UX 개선
 Top3 컴포넌트 레이아웃 스타일링 개선
