+++ apps/web/app/error.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Button from '@/shared/ui/Button';
+import Button from '@/shared/ui/Button/Button';
 import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
