+++ apps/web/src/shared/ui/Link/MobileWirtePostLink.tsx
